Updated setup script to make it safe to rerun.

+++ scripts/setup.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 # Make sure only root can run our script
 if [[ $EUID -ne 0 ]]; then
@@ -9,18 +9,22 @@ fi
 # add redis package to apt
 echo "Adding Redis package to apt..."
 pushd /etc/apt/sources.list.d/
-touch ./dotdeb.org.list
-echo "deb http://packages.dotdeb.org wheezy all" >> ./dotdeb.org.list
-echo "deb-src http://packages.dotdeb.org wheezy all" >> ./dotdeb.org.list
-wget -q -O - http://www.dotdeb.org/dotdeb.gpg | apt-key add -
+if [ ! -f ./dotdeb.org.list ]; then
+  touch ./dotdeb.org.list
+  echo "deb http://packages.dotdeb.org wheezy all" >> ./dotdeb.org.list
+  echo "deb-src http://packages.dotdeb.org wheezy all" >> ./dotdeb.org.list
+  wget -q -O - http://www.dotdeb.org/dotdeb.gpg | apt-key add -
+fi
 popd
 
 # add rabbit package to apt
 echo "Adding Rabbit package to apt..."
 pushd /etc/apt/sources.list.d/
-touch ./rabbitmq.com.list
-echo "deb http://www.rabbitmq.com/debian/ testing main" >> ./rabbitmq.com.list
-wget -q -O - http://www.rabbitmq.com/rabbitmq-signing-key-public.asc | apt-key add -
+if [ ! -f ./rabbitmq.com.list ]; then
+  touch ./rabbitmq.com.list
+  echo "deb http://www.rabbitmq.com/debian/ testing main" >> ./rabbitmq.com.list
+  wget -q -O - http://www.rabbitmq.com/rabbitmq-signing-key-public.asc | apt-key add -
+fi
 popd
 
 # update apt and install packages
@@ -29,12 +33,13 @@ apt-get install -y build-essential openssl htop curl make vim nautilus-open-term
 
 # enable rabbit admin console on port 15672 and restart
 rabbitmq-plugins enable rabbitmq_management
-rabbitmqctl stop_app
-rabbitmqctl start_app
+service rabbitmq-server restart
 
-# install nvm (node version manager)
-wget -qO- https://raw.github.com/creationix/nvm/master/install.sh | sh
-source ~/.nvm/nvm.sh && echo -e "\n. ~/.nvm/nvm.sh" >> ~/.bashrc
+if [ ! -d ~/.nvm ]; then
+  # install nvm (node version manager)
+  wget -qO- https://raw.github.com/creationix/nvm/master/install.sh | sh
+  source ~/.nvm/nvm.sh && echo -e "\n. ~/.nvm/nvm.sh" >> ~/.bashrc
+fi
 
 # install the newest nodejs 0.10.x and set it as the default nodejs
 nvm install 0.10 && nvm alias default 0.10
